fix(genproj): extract shared api logic to fix code duplication
Extracted the error handling, auth token building, and project context
creation logic that was previously duplicated between `POST /api/v1/genproj`
and `POST /projects/genproj/api/generate` into a new shared utility module
`$lib/server/genproj-api-utils.js`.

This resolves the SonarCloud code duplication CI failure.
